Do not incorrectly add line separators for non-synthetic nodes when emitting node list
As of 3c32f6e154ead6749b76ec9c19cbfdd2acad97d6, a line separator is
added between nodes if the nodes are not synthetic and on separate
lines. This it push s wrong and previously only happened if the non-synthetic
nodes were on different lines but had the same parent.

Fixes #44068.