Prototype implementation of Source Maps v4
Source Maps v4 standard proposal is an addendum to source maps to
improve post-hoc debugging and analysis of call stacks. This proposal is
being put forward in
[https://github.com/MicrosoftEdge/MSEdgeExplainers/pull/538](Microsoft Edge)
and is designed to enable an analysis of stack traces to reconstruct an
unminified stack trace.

To do so, two additional fields are added to the source maps output:
 - `scopeNames`
 - `scopes`

The latter field behaves similarly to the existing `mappings` field. The
former list are names that should be displayed for a given function once
it is determined to be responsible. The format of the `scopes` field is
a Base64 VLQ with 6 fields, which are, in order:
 - Source file index
 - Source line (relative offset)
 - Source character (relative offset)
 - End line (relative offset, should always be positive)
 - End character (relative offset)
 - Index into `scopeNames`

(This first commit does not implement the VLQ encoding; rather, all
number values are directly passed in and are neither relative nor are
they encoded as VLQs).