Add undefined to parameter's name type
Only parameters to JSDocFunctionTypes may not have names, but these
parameters cause crashes from time to time, most recently #47606.

This PR is NOT in a mergeable state -- having fixed all the compile
errors, I don't actually think it's worthwhile. A lot of the difficulty
comes from deciding whether to handle an undefined name instead of just
asserting that it's defined. Any piece of code that only handles
function *values* doesn't need to handle undefined parameters. However,
it's difficult to know when this is the case.