JSDoc comment parsing supports multiline backticks
Previously, backticks in JSDoc ended at the end of a line. This PR
changes that so that backticks continue parsing until another backtick.

Note that triple backticks and single backticks are treated the same
way. The change is already very clunky, and I haven't thought about how
much clunkier it would be to track backtick history in the parser.