Handle changes to inferred projects during FAR
In some scenarios (see the new test for an example), FAR may search an inferred project and then subsequently trigger the loading of a configured project containing (some of) the same files.  When the configured project is loaded, the files are removed from the inferred project and added to the configured project, potentially _invalidating the results already computed for the inferred project_.  This change introduces a cleanup pass that removes results from `dirty` projects if the corresponding files are no longer in the project.

Caught by #50679