Adds optional parameter to createPrinter (in emitter.ts) to allow specifying the indentation either as a number of spaces or a sequence of 1+ characters (be that tab(s) or any other character(s) of choice).
This pull request addresses open issue https://github.com/microsoft/TypeScript/issues/2306.

How this works?  Currently the indentation is hardcoded in the utilities.ts (called indentStrings) as a cached array of strings where each string in the array is the next step of indentation to be used... position 0 = "", 1 = "    ", 2 = "        ", etc.  ... and it cannot be changed.  The modifications in this pull request includes a method that will 'rebuild' the content of that array using the information provided by a new PrinterOptions property called 'indentation'.  Note:  the array is the same instance... just the content changes.

If 'indentation' is not supplied (the default (and also fully backwards compatiable) or is not a string or number) then the cached array is left as is.
If 'indentation' is a number between 0 to 9 (negative numbers or numbers higher than 9 are treated as a default of 4) then the cached array content is rebuilt with strings incremented in each subsequent position with the supplied number of spaces...   e.g.   3 = ["", "   ", "      ", ...]
if 'indentation' is a string then the cached array content is rebuilt with that string as the indentation step... "\t" = ["", "\t", "\t\t", ...]

Special note: The current implementation of the indentation is a 'static' cached array (all createPrinter instances use the same array).  It is possibly to see unexpected results when 2 or more instances of a printer are created in parallel (and the indentation is different).  The latest createPrinter call will dictate the size and content of each indentation step at the time of the call.  This can only be fixed by moving the indentation cache into the 'printer' instance and is beyond the scope of this pull request.