Fix APISample tests to not depend on the existence of the ts namespace
The way we pull in the types for typescript in these tests only works
when we include typescriptServices.d.ts into the project, which declares
a global namespace, which then is reexported as the typescript package.

This reexport is already what the current typescript.d.ts does, but now
that we don't have the namespace at all, we can't do this. Instead,
simulate the typescript package directly and point it at the correct
d.ts file.

This is a good chagne regardless of the module transform, and could be
cherry-picked back to main early.