marking options as publicly available
There's several fields in CompilerOptions that are marked internal but are now publicly supported. It would be great to be able to get at them in a typesafe way.

I left plugin as internal because I don't have a ton of context on that option. Feel free to hand this over to someone with way more context.