Remove generation of protocol.d.ts
This file is unreferenced, and just a subset of tsserverlibrary.d.ts.
What I've heard is that this was intended to be used to detect protocol
chagnes, but, no test actually references it, and we have separate
infrastructure in place to ensure that protocol changes get the
attention they need.